# spec/openapi.yaml
openapi: '3.0.2'
info:
  title: Sbullasy HTTP API Specification
  # summary:
  description: This is the specification of Sbullasy HTTP API.
  # termsOfService:
  # contact:
  # license:
  version: '0.1.0'

# jsonSchemaDialect:
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        # enum:
        default: sbullasy.mgn901.com
        # description:
paths:
...
  /users/me/challenges/{requestId}:
    put:
      tags:
        - users
      description: Answer to a challenge.
      parameters:
        - $ref: '#/components/parameters/requestId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: './http-api-json-schemas/challenge-for-payload.yaml'
      responses:
        204:
          description: Indicates the challenge has been admitted.
        401:
          $ref: '#/components/responses/e401NoToken'
        403:
          $ref: '#/components/responses/e403WrongChallenge'
...
components:
...
  responses:
    e401NoToken:
      description: Indicates the client doesn't have any tokens.
      content:
        application/json:
          schema:
            $ref: './http-api-json-schemas/error.yaml'
...
    e403WrongChallenge:
      description: Indicates the client sent a challenge with wrong answer.
      content:
        application/json:
          schema:
            $ref: './http-api-json-schemas/error.yaml'
...
  parameters:
...
    requestId:
      name: requestId
      in: path
      schema:
        type: string
      required: true
...
  securitySchemes:
    cookie_auth:
      type: apiKey
      # description:
      name: token
      in: cookie
    bearer_auth:
      type: http
      scheme: bearer
  # links:
  # callbacks:
  # pathItems:
# security:

# tags:

# externalDocs: